move client work to client folder
