+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules
 server/node_modules
 dist
+.parcel-cache
