+++ package.json
@@ -6,8 +6,10 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "server": "npm run server -w @geckochat/server",
+    "client": "npm run client -w @geckochat/client",
     "server:lint": "npm run lint -w @geckochat/server",
-    "start": "parcel",
+    "client:lint": "npm run lint -w @geckochat/client",
+    "pre-commit": "lint-staged",
     "prepare": "husky install"
   },
   "repository": {
@@ -21,20 +23,24 @@
   },
   "homepage": "https://github.com/chingu-voyages/v41-geckos-team-12#readme",
   "devDependencies": {
-    "@types/react-dom": "^18.0.6",
     "eslint": "^8.25.0",
     "husky": "^8.0.1",
     "lint-staged": "^13.0.3",
-    "parcel": "^2.7.0",
     "prettier": "2.7.1",
-    "process": "^0.11.10",
     "typescript": "^4.8.4"
   },
   "workspaces": [
-    "server"
+    "server",
+    "client"
   ],
-  "dependencies": {
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+  "lint-staged": {
+    "server/*.js": [
+      "prettier --write .",
+      "eslint . --fix"
+    ],
+    "client/*.js": [
+      "prettier --write .",
+      "eslint . --fix"
+    ]
   }
 }
