--- src/App.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function App() {
-  return (
-    <div>
-      <h1>Chat app</h1>
-    </div>
-  )
-}
-
-export default App
